refactor(BulkOp): refactor bulk op code to remove redundancies
