+++ lib/bulk/common.js
@@ -2,12 +2,15 @@
 
 const Long = require('mongodb-core').BSON.Long;
 const MongoError = require('mongodb-core').MongoError;
+const ObjectID = require('mongodb-core').BSON.ObjectID;
+const BSON = require('mongodb-core').BSON;
+const MongoWriteConcernError = require('mongodb-core').MongoWriteConcernError;
 const toError = require('../utils').toError;
 const handleCallback = require('../utils').handleCallback;
 const applyRetryableWrites = require('../utils').applyRetryableWrites;
 const applyWriteConcern = require('../utils').applyWriteConcern;
-const ObjectID = require('mongodb-core').BSON.ObjectID;
-const BSON = require('mongodb-core').BSON;
+const executeOperation = require('../utils').executeOperation;
+const isPromiseLike = require('../utils').isPromiseLike;
 
 // Error codes
 const UNKNOWN_ERROR = 8;
@@ -352,7 +355,7 @@ class WriteError {
  * Merges results into shared data structure
  * @ignore
  */
-function mergeBatchResults(ordered, batch, bulkResult, err, result) {
+function mergeBatchResults(batch, bulkResult, err, result) {
   // If we have an error set the result to be the err object
   if (err) {
     result = err;
@@ -484,6 +487,41 @@ function mergeBatchResults(ordered, batch, bulkResult, err, result) {
   }
 }
 
+function executeCommands(bulkOperation, options, callback) {
+  if (bulkOperation.s.batches.length === 0) {
+    return handleCallback(callback, null, new BulkWriteResult(bulkOperation.s.bulkResult));
+  }
+
+  const batch = bulkOperation.s.batches.shift();
+
+  function resultHandler(err, result) {
+    // Error is a driver related error not a bulk op error, terminate
+    if (((err && err.driver) || (err && err.message)) && !(err instanceof MongoWriteConcernError)) {
+      return handleCallback(callback, err);
+    }
+
+    // If we have and error
+    if (err) err.ok = 0;
+    if (err instanceof MongoWriteConcernError) {
+      return handleMongoWriteConcernError(batch, bulkOperation.s.bulkResult, err, callback);
+    }
+
+    // Merge the results together
+    const writeResult = new BulkWriteResult(bulkOperation.s.bulkResult);
+    const mergeResult = mergeBatchResults(batch, bulkOperation.s.bulkResult, err, result);
+    if (mergeResult != null) {
+      return handleCallback(callback, null, writeResult);
+    }
+
+    if (bulkOperation.handleWriteError(callback, writeResult)) return;
+
+    // Execute the next command in line
+    executeCommands(bulkOperation, options, callback);
+  }
+
+  bulkOperation.finalOptionsHandler({ options, batch, resultHandler }, callback);
+}
+
 /**
  * handles write concern error
  *
@@ -493,8 +531,8 @@ function mergeBatchResults(ordered, batch, bulkResult, err, result) {
  * @param {WriteConcernError} err
  * @param {function} callback
  */
-function handleMongoWriteConcernError(batch, bulkResult, ordered, err, callback) {
-  mergeBatchResults(ordered, batch, bulkResult, null, err.result);
+function handleMongoWriteConcernError(batch, bulkResult, err, callback) {
+  mergeBatchResults(batch, bulkResult, null, err.result);
 
   const wrappedWriteConcernError = new WriteConcernError({
     errmsg: err.result.writeConcernError.errmsg,
@@ -955,6 +993,18 @@ class BulkOperationBase {
     return { options, callback };
   }
 
+  execute(_writeConcern, options, callback) {
+    const ret = this.bulkExecute(_writeConcern, options, callback);
+    if (!ret || isPromiseLike(ret)) {
+      return ret;
+    }
+
+    options = ret.options;
+    callback = ret.callback;
+
+    return executeOperation(this.s.topology, executeCommands, [this, options, callback]);
+  }
+
   /**
    * Handles final options before executing command
    *
@@ -1038,11 +1088,7 @@ class BulkOperationBase {
       // Force top level error
       err.ok = 0;
       // Merge top level error and return
-      handleCallback(
-        callback,
-        null,
-        mergeBatchResults(false, config.batch, this.s.bulkResult, err, null)
-      );
+      handleCallback(callback, null, mergeBatchResults(config.batch, this.s.bulkResult, err, null));
     }
   }
