+++ lib/bulk/unordered.js
@@ -2,17 +2,10 @@
 
 const common = require('./common');
 const BulkOperationBase = common.BulkOperationBase;
+const Batch = common.Batch;
+const bson = common.bson;
 const utils = require('../utils');
 const toError = utils.toError;
-const handleCallback = utils.handleCallback;
-const BulkWriteResult = common.BulkWriteResult;
-const Batch = common.Batch;
-const mergeBatchResults = common.mergeBatchResults;
-const executeOperation = utils.executeOperation;
-const MongoWriteConcernError = require('mongodb-core').MongoWriteConcernError;
-const handleMongoWriteConcernError = require('./common').handleMongoWriteConcernError;
-const bson = common.bson;
-const isPromiseLike = require('../utils').isPromiseLike;
 
 /**
  * Add to internal list of Operations
@@ -130,60 +123,6 @@ class UnorderedBulkOperation extends BulkOperationBase {
    * @throws {MongoError} Throws error if the bulk object does not have any operations
    * @return {Promise} returns Promise if no callback passed
    */
-  execute(_writeConcern, options, callback) {
-    const ret = this.bulkExecute(_writeConcern, options, callback);
-    if (!ret || isPromiseLike(ret)) {
-      return ret;
-    }
-
-    options = ret.options;
-    callback = ret.callback;
-
-    return executeOperation(this.s.topology, executeCommands, [this, options, callback]);
-  }
-}
-
-/**
- * Execute next write command in a chain
- *
- * @param {OrderedBulkOperation} bulkOperation
- * @param {object} options
- * @param {function} callback
- */
-function executeCommands(bulkOperation, options, callback) {
-  if (bulkOperation.s.batches.length === 0) {
-    return handleCallback(callback, null, new BulkWriteResult(bulkOperation.s.bulkResult));
-  }
-
-  // Ordered execution of the command
-  const batch = bulkOperation.s.batches.shift();
-
-  function resultHandler(err, result) {
-    // Error is a driver related error not a bulk op error, terminate
-    if (((err && err.driver) || (err && err.message)) && !(err instanceof MongoWriteConcernError)) {
-      return handleCallback(callback, err);
-    }
-
-    // If we have and error
-    if (err) err.ok = 0;
-    if (err instanceof MongoWriteConcernError) {
-      return handleMongoWriteConcernError(batch, bulkOperation.s.bulkResult, true, err, callback);
-    }
-
-    // Merge the results together
-    const writeResult = new BulkWriteResult(bulkOperation.s.bulkResult);
-    const mergeResult = mergeBatchResults(true, batch, bulkOperation.s.bulkResult, err, result);
-    if (mergeResult != null) {
-      return handleCallback(callback, null, writeResult);
-    }
-
-    if (bulkOperation.handleWriteError(callback, writeResult)) return;
-
-    // Execute the next command in line
-    executeCommands(bulkOperation, options, callback);
-  }
-
-  bulkOperation.finalOptionsHandler({ options, batch, resultHandler }, callback);
 }
 
 /**
